add linux 64 fmod lib, fix mp3 frame detection code, which was detected on 64 bit linux, but is an error on every platform

+++ trunk/wxMusik/src/Classes/Library/mp3tech.h
@@ -50,7 +50,7 @@ enum VBR_REPORT { VBR_VARIABLE, VBR_AVERAGE, VBR_MEDIAN };
 enum SCANTYPE { SCAN_NONE, SCAN_QUICK, SCAN_FULL };
 
 typedef struct {
-	unsigned long	sync;
+	unsigned int	sync;
 	unsigned int	version;
 	unsigned int	layer;
 	unsigned int	crc;
